refactor: extract oss llm client
