+++ docs/PLANS.md
@@ -73,6 +73,7 @@
 │   ├── OSS/
 │   │   ├── Open_AI_OSS.py          # 챗봇 메인 오케스트레이션
 │   │   ├── service.py              # 챗봇 서비스 연결부
+│   │   ├── llm_client.py           # OSS/OpenAI 호환 LLM client 생성 및 호출
 │   │   ├── tools.py                # 비용 우선 deterministic tool routing
 │   │   ├── formatter.py            # 응답 포맷팅
 │   │   ├── modes.py                # 챗봇 모드 정의
@@ -148,6 +149,7 @@
 
 - `LLM/OSS/Open_AI_OSS.py`
 - `LLM/OSS/service.py`
+- `LLM/OSS/llm_client.py`
 - `LLM/OSS/tools.py`
 - `LLM/OSS/formatter.py`
 - `LLM/OSS/modes.py`
